Use system xdg-shell protocol when available.
This change matters when linking as a static library or meson subproject
to the project that is already building xdg-shell protocol of a
different version. In this case linker would prefer symbols from the
main project.
If both the main project and the subproject are using the same xdg-shell
protocol definition from the system wayland-protocols, the chance of
things going wrong is slightly lesser