Normalize gboolean values for anchors.
The code in this file assumes that all true gbooleans are equal. That's
not correct and it's acceptable, albeit discouraged, to assign any int
value to gboolean. Of course in this case any strict equality checks in
this file could fail.

Fixes auto exclusive zone when passing values other than TRUE or FALSE
to `gtk_layer_set_anchor`.