gtk-wayland: handle xdg_wm_base::ping
This seems to be necessary on KWin, otherwise regular xdg-toplevels would be detected as "unresponsible". Note: I assume that GTK already does this, but since we connect to xdg_wm_base separately, this seems required here as well (I don't know if KWin picks arbitrarily one xdg_wm_base object to send pings to or sends it to all of them though).